修正 Windows 平台下编译错误问题

+++ include/quorum/majority.h
@@ -90,7 +90,7 @@ class majority_config {
 
     for (std::size_t i = 0; i < info.size(); i++) {
       if ((i > 0) && (info[i - 1].index < info[i].index)) {
-        info[i].bar = i;
+        info[i].bar = static_cast<int>(i);
       }
     }
